test: wip workflow tests
