+++ packages/scripts/src/commands/workflow/run.ts
@@ -2,37 +2,14 @@
 
 import boxen from "boxen";
 import chalk from "chalk";
-import path from "path";
 import { Command } from "commander";
+import path from "path";
+
 import { IDeploymentWorkflows, IWorkflow, WORKFLOW_DEFAULTS } from "workflows";
 import ALL_TASKS from "../../tasks";
 import { Logger, logProgramHelp, pad, promptOptions } from "../../utils";
 import { ActiveDeployment } from "../deployment/get";
 import type { IDeploymentConfigJson } from "../deployment/common";
-
-const program = new Command("run");
-
-interface IProgramOptions {
-  parent?: string;
-}
-
-/***************************************************************************************
- * CLI
- * @example
- *************************************************************************************/
-export default program
-  .description("Run a workflow")
-  .argument("[name]", "Name of workflow to run")
-  .allowUnknownOption()
-  .helpOption("--helpIgnored", "will show help from child workflow instead of this")
-  .option("-p --parent <string>", "Name of parent workflow triggered by")
-  .action(async (name, opts: IProgramOptions) => {
-    // pass any additional args after [name] positional argument
-    const args = program.args.slice(1);
-    const runner = WorkflowRunner;
-    await runner.init();
-    return runner.run({ ...opts, name, args });
-  });
 
 /***************************************************************************
  *  Main Methods
@@ -96,9 +73,9 @@ export class WorkflowRunnerClass {
    **/
   private prepareWorkflow(name: string, args: string[] = []) {
     // include manual help logging as default ignored (so can pass to child command)
-    if (["--help", "-h"].includes(name)) {
-      logProgramHelp(program);
-    }
+    // if (["--help", "-h"].includes(name)) {
+    //   logProgramHelp(program);
+    // }
     let workflow = this.workflows[name];
     // Ensure workflow exists
     if (!workflow) {
@@ -189,24 +166,3 @@ export class WorkflowRunnerClass {
 
 /** Create single instance that can be shared (for dependency injection) */
 export const WorkflowRunner = new WorkflowRunnerClass();
-
-/***************************************************************************
- *  Logging and Error Handling
- **************************************************************************/
-const cleanup = () => {
-  console.log("Cleaning up.");
-};
-const handleExit = () => {
-  cleanup();
-  console.log("Exiting without error.");
-  process.exit();
-};
-const handleError = (e) => {
-  console.error("ERROR! An error was encountered while executing");
-  console.error(e);
-  cleanup();
-  console.log("Exiting with error.");
-  process.exit(1);
-};
-process.on("SIGINT", handleExit);
-process.on("uncaughtException", handleError);
