+++ packages/scripts/src/commands/workflow/workflow.spec.ts
@@ -1,0 +1,10 @@
+import { WorkflowRunnerClass } from "./run";
+
+describe("Workflow", () => {
+  fit("Runs a workflow", () => {
+    console.log({ WorkflowRunnerClass });
+    // const runner = new WorkflowRunnerClass();
+    console.log("hello runner");
+    expect(true).toEqual(false);
+  });
+});
